+++ src/components/creator/Editor/SceneEdition/Grid/SceneCell.tsx
@@ -1,7 +1,7 @@
 import styles from "./grid.module.css"
 import { useContext } from "react"
 import { CreatorContext } from "../../CreatorContext"
-import { ACTOR, INITIAL_ROW, INITIAL_COL, OBSTACLE, EMPTY, setActorAtPosition } from "../SceneEdition"
+import { ACTOR, INITIAL_ROW, INITIAL_COL, OBSTACLE, EMPTY, setActorAtPosition, actorPosition, hasActor } from "../gridUtils"
 import { SceneMap, SceneType } from "../../../../serializedChallenge"
 
 type CellProps = {
@@ -24,11 +24,11 @@ export const SceneCell: React.FC<CellProps> = (props) => {
 
     const objectsInCell = props.content.split('&').filter(o => o !== '-')
 
+    const currentContent = props.content
+
     const objectStyle = (object: string) => styles[`img-${object}`] || styles['img-default']
 
     const hasMultipleObjects = (cellObjects = objectsInCell): boolean => cellObjects.length > 1
-
-    const hasActor = (cell = props.content): boolean => cell.split('&').includes(ACTOR)
 
     const isInitialCell: boolean = props.position.row === INITIAL_ROW && props.position.column === INITIAL_COL
 
@@ -48,38 +48,32 @@ export const SceneCell: React.FC<CellProps> = (props) => {
     const handleActor = (): SceneMap => {
         return setActorAtPosition(deletedActorMap(), props.position.row, props.position.column)
     }
-
-    const deletedActorMap = () => {
-        let prevPos = actorPosition()
-        if( prevPos.row > -1 && prevPos.column > -1 )
-            {
-            let cellObjects = map[prevPos.row][prevPos.column].split('&')
-            map[prevPos.row][prevPos.column] = hasMultipleObjects(cellObjects) ? cellObjects[1] : EMPTY
-            }
-        return map
+    
+    const handleEraser = (): SceneMap => {
+        if (hasActor(currentContent) && isInitialCell && !hasMultipleObjects()) return map; // We can't erase actor on the initial cell
+        if (hasActor(currentContent) && !hasMultipleObjects()) map = setActorAtPosition(map)
+        return mapWithNewCellContent(hasMultipleObjects() ? ACTOR : EMPTY)
     }
 
-    const handleEraser = (): SceneMap => {
-        if (hasActor() && isInitialCell && !hasMultipleObjects()) return map; // We can't erase actor on the initial cell
-        if (hasActor() && !hasMultipleObjects()) map = setActorAtPosition(map)
-        return mapWithNewCellContent(hasMultipleObjects() ? ACTOR : EMPTY)
-    } 
-
     const handlePrize = (): SceneMap => {
-        return mapWithNewCellContent(hasActor() ? ACTOR + '&' + selectedTool : selectedTool)
+        return mapWithNewCellContent(hasActor(currentContent) ? ACTOR + '&' + selectedTool : selectedTool)
     }
 
     const handleObstacle = (): SceneMap => {
-        if (hasActor() && isInitialCell) return map; // We can't replace actor on the initial cell
-        if (hasActor() && !isInitialCell) map = setActorAtPosition(map)
+        if (hasActor(currentContent) && isInitialCell) return map; // We can't replace actor on the initial cell
+        if (hasActor(currentContent) && !isInitialCell) map = setActorAtPosition(map)
         return mapWithNewCellContent(selectedTool)
-    } 
-
-    const actorPosition = (): Position => {
-        let row = map.findIndex(row => row.some(hasActor))
-        let column = (row > -1 ? map[row].findIndex(hasActor):-1)
-        return { row, column }
     }
+    const deletedActorMap = () => {
+        let prevPos = actorPosition(map)
+        if (!prevPos) return map //if there is no actor in the map
+
+        let cellObjects = map[prevPos.row][prevPos.column].split('&')
+        map[prevPos.row][prevPos.column] = hasMultipleObjects(cellObjects) ? cellObjects[1] : EMPTY
+
+        return map
+    }
+
 
     const mapWithNewCellContent = (content: string, position: Position = props.position): SceneMap => {
         map[position.row][position.column] = content
